# asset-log/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AssetClass": {
        "enum": [
          "equity",
          "etf",
          "mutual_fund",
          "bond",
          "cash",
          "other"
        ],
        "type": "string"
      },
      "AssetResponse": {
        "properties": {
          "asset_class": {
            "$ref": "#/components/schemas/AssetClass"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "currency": {
            "example": "JPY",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "example": "トヨタ自動車",
            "type": "string"
          },
          "price_unit": {
            "description": "価格の単位（投信は10000、それ以外は1）。文字列で表現される",
            "example": "1",
            "type": "string"
          },
          "symbol": {
            "description": "ティッカーや証券コード",
            "example": "7203",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "symbol",
          "name",
          "asset_class",
          "currency",
          "price_unit",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
...
      "CreateAssetRequest": {
        "properties": {
          "asset_class": {
            "$ref": "#/components/schemas/AssetClass"
          },
          "currency": {
            "description": "ISO 4217 の3文字。小文字で送っても大文字に正規化される",
            "example": "JPY",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "price_unit": {
            "description": "未指定なら資産クラスの既定値（投信は10000、他は1）",
            "example": "10000",
            "type": [
              "string",
              "null"
            ]
          },
          "symbol": {
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "name",
          "asset_class"
        ],
        "type": "object"
      },
...
      "Credentials": {
        "description": "ログイン・登録の認証情報",
        "properties": {
          "email": {
            "description": "メールアドレス（大文字小文字は区別しない）",
            "example": "user@example.com",
            "type": "string"
          },
          "password": {
            "description": "パスワード（登録時は12文字以上）",
            "example": "correct-horse-battery-staple",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
      "FieldError": {
        "properties": {
          "field": {
            "description": "エラーが発生した入力項目名",
            "example": "quantity",
            "type": "string"
          },
          "message": {
            "description": "項目ごとのエラー内容",
            "example": "数量は正の数である必要があります",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "type": "object"
      },
...
      "MeResponse": {
        "description": "認証済みユーザーの情報",
        "properties": {
          "user_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "user_id"
        ],
        "type": "object"
      },
      "PatchAssetRequest": {
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "price_unit": {
            "type": [
              "string",
              "null"
            ]
          },
          "symbol": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "ProblemDetails": {
        "description": "RFC 9457 準拠のエラーレスポンス（`application/problem+json`）",
        "properties": {
          "detail": {
            "description": "人間向けの詳細メッセージ",
            "example": "口座が見つかりません",
            "type": "string"
          },
          "errors": {
            "description": "項目単位のエラー（422のみ。空の場合はフィールドごと省略される）",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "status": {
            "description": "HTTPステータスコード",
            "example": 404,
            "format": "int32",
            "maximum": 599,
            "minimum": 100,
            "type": "integer"
          },
          "title": {
            "description": "エラー種別の短い説明",
            "example": "Not Found",
            "type": "string"
          },
          "trace_id": {
            "description": "サーバーログと突き合わせるための識別子",
            "format": "uuid",
            "type": "string"
          },
          "type": {
            "description": "エラー種別の安定した識別子。クライアントはこの値で分岐できる",
            "example": "/errors/not-found",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "trace_id"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string"
          },
          "expires_in": {
            "format": "int64",
            "type": "integer"
          },
          "token_type": {
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_in"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "jobToken": {
        "description": "バッチ実行用の固定トークン",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "NISA/iDeCo を含む資産・取引の管理API",
    "license": {
      "name": ""
    },
    "title": "asset-log API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/assets": {
      "get": {
        "operationId": "list_assets",
        "parameters": [
          {
            "description": "シンボルまたは名称の部分一致検索",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssetResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "銘柄の一覧"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "assets"
        ]
      },
      "post": {
        "operationId": "create_asset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            },
            "description": "銘柄を作成した"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "このシンボルは既に登録されている"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "シンボル・名称が空、通貨コードの形式不正、価格単位が0以下"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "assets"
        ]
      }
    },
    "/assets/{id}": {
      "get": {
        "operationId": "get_asset",
        "parameters": [
          {
            "description": "銘柄ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            },
            "description": "銘柄の詳細"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "銘柄が存在しない"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "assets"
        ]
      },
      "patch": {
        "operationId": "update_asset",
        "parameters": [
          {
            "description": "銘柄ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchAssetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            },
            "description": "更新後の銘柄"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "更新する項目が指定されていない"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "銘柄が存在しない"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "シンボルが他の銘柄と重複"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "シンボル・名称が空、価格単位が0以下"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "assets"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "ログイン成功"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "メールアドレスまたはパスワードが違う"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "登録成功。アクセストークンを返す"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "このメールアドレスは既に登録されている"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "入力値が要件を満たしていない"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/me": {
      "get": {
        "operationId": "me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "認証済みユーザーの情報"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "保護API の動作確認用。タスク#5以降のCRUDも同じ形で書ける。",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "ローカル開発",
      "url": "http://localhost:8080"
    }
  ],
...
}